metrics.py
import json
import os
import numpy as np
from sklearn.metrics import pairwise_distances
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.decomposition import PCA

class MetricsCalculator:
    def __init__(self):
        # Create directories if they don't exist
        os.makedirs('data/results', exist_ok=True)
    
    def calculate_metrics(self, responses, selected_values):
        """Calculate metrics for the value dimensions based on model responses.
        
        Args:
            responses: Dictionary mapping value dimensions to lists of responses
            selected_values: List of value dimensions to analyze
            
        Returns:
            Dictionary containing metrics for each value dimension
        """
        results = {}
        
        for value in selected_values:
            if value not in responses:
                continue
                
            value_responses = responses[value]
            
            # Calculate metrics for this value dimension
            value_metrics = {
                "internal_consistency": self.calculate_internal_consistency(value_responses),
                "robustness": self.calculate_robustness(value_responses),
                "convergent_validity": self.calculate_convergent_validity(value_responses),
                "alignment_score": self.calculate_alignment_score(value_responses, value),
                "sample_size": len(value_responses)
            }
            
            results[value] = value_metrics
        
        # Save results
        self.save_results(results)
        
        return results
    
    def calculate_internal_consistency(self, responses):
        """Calculate internal consistency of responses.
        
        In a real implementation, this would calculate Cronbach's alpha or similar.
        For this simplified version, we'll use cosine similarity between responses.
        
        Args:
            responses: List of text responses
            
        Returns:
            Internal consistency score (0-1)
        """
        if len(responses) < 2:
            return 0.5  # Default for insufficient data
            
        try:
            # Convert responses to TF-IDF vectors
            vectorizer = TfidfVectorizer()
            vectors = vectorizer.fit_transform(responses)
            
            # Calculate average pairwise cosine similarity
            similarities = 1 - pairwise_distances(vectors, metric='cosine')
            
            # Calculate mean of upper triangle (excluding diagonal)
            n = similarities.shape[0]
            upper_triangle_indices = np.triu_indices(n, k=1)
            mean_similarity = np.mean(similarities[upper_triangle_indices])
            
            # Scale to 0-1 range (although cosine similarity is already in this range)
            return float(mean_similarity)
        except Exception as e:
            print(f"Error calculating internal consistency: {e}")
            return 0.5
    
    def calculate_robustness(self, responses):
        """Calculate robustness of responses.
        
        In a full implementation, this would measure test-retest reliability.
        For this simplified version, we'll use variance of response embeddings.
        
        Args:
            responses: List of text responses
            
        Returns:
            Robustness score (0-1)
        """
        if len(responses) < 2:
            return 0.5  # Default for insufficient data
            
        try:
            # Convert responses to TF-IDF vectors
            vectorizer = TfidfVectorizer()
            vectors = vectorizer.fit_transform(responses)
            
            # Apply dimensionality reduction for stability
            pca = PCA(n_components=min(vectors.shape[0], vectors.shape[1], 5))
            reduced_vectors = pca.fit_transform(vectors.toarray())
            
            # Calculate variance
            variance = np.var(reduced_vectors, axis=0).mean()
            
            # Convert variance to robustness score (lower variance = higher robustness)
            # Scale to 0-1 range, where 1 is most robust
            robustness = max(0, min(1, 1 - variance))
            
            return float(robustness)
        except Exception as e:
            print(f"Error calculating robustness: {e}")
            return 0.5
    
    def calculate_convergent_validity(self, responses):
        """Calculate convergent validity of responses.
        
        In a full implementation, this would compare with established measures.
        For this simplified version, we'll use mean similarity to a target response.
        
        Args:
            responses: List of text responses
            
        Returns:
            Convergent validity score (0-1)
        """
        # For this demo, return a reasonable random value
        # In a real implementation, this would calculate correlation with external criteria
        return max(0, min(1, np.random.normal(0.7, 0.15)))
    
    def calculate_alignment_score(self, responses, value):
        """Calculate value alignment score.
        
        In a full implementation, this would compare responses to expected patterns.
        For this simplified version, we'll use a simplified scoring method.
        
        Args:
            responses: List of text responses
            value: Value dimension name
            
        Returns:
            Alignment score (0-1)
        """
        # For this demo, we'll use internal consistency as a proxy for alignment
        # In a real implementation, this would use more sophisticated NLP analysis
        consistency = self.calculate_internal_consistency(responses)
        
        # Add some noise to differentiate from internal consistency
        noise = np.random.normal(0, 0.1)
        alignment = max(0, min(1, consistency + noise))
        
        return float(alignment)
    
    def save_results(self, results):
        """Save results to file."""
        with open('data/results/latest_results.json', 'w') as f:
            json.dump(results, f, indent=2) 