requirements.txt
Flask==2.3.3
openai>=1.0.0
anthropic>=0.25.0
google-generativeai>=0.3.0
plotly>=5.15.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
Jinja2>=3.1.2
Werkzeug>=2.3.7
click>=8.1.0
pyyaml>=6.0.1
anthropic
google-generativeai
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.20
alembic==1.12.0
ruamel.yaml
opencv-python